requirements.txt
Mastodon.py==1.3.1
markovify==0.7.1
beautifulsoup4==4.6.3
click=6.7